pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "android-mcp"
version = "0.1.0"
description = "An MCP server for android automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pure-python-adb>=0.3.0.dev0",
    "PyYAML>=6.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0", "pytest-mock>=3.12.0", "pytest-cov>=4.0.0"]

[tool.setuptools]
py-modules = ["server", "adbdevicemanager", "exceptions", "xml_utils", "retry"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "--disable-warnings", "-v"]
